Update codebase with new changes
This commit includes various updates and modifications to the existing codebase as per the latest requirements.
[skip gpt_engineer]
